backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
aiohttp==3.10.5
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
google-generativeai==0.8.2
composio-core==0.6.8
composio-google==0.6.8
sse-starlette==2.1.3
python-multipart==0.0.12
requests==2.32.3
rich==13.8.1
loguru==0.7.2
